% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronaaus-package.R
\docType{package}
\name{coronaaus-package}
\alias{coronaaus}
\alias{coronaaus-package}
\title{coronaaus: A Tool to Assist with the Analysis of Coronavirus in Australia}
\description{
This package has a variety of functions that enable users to explore coronavirus data in Australia via the application. Most of the functions are designed to streamline the application, which can be launched via the `launch_app` function.
}
\author{
\strong{Maintainer}: Emily Sheehan \email{eshe0012@student.monash.edu}

}
\keyword{internal}
