% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_title.R
\name{covid_title}
\alias{covid_title}
\title{Produce a title for the Map and Table outputs}
\usage{
covid_title(type)
}
\arguments{
\item{type}{The type of output as a character vector (must be a map or table).}
}
\value{
The title of the output
}
\description{
This function automatically generates a title for the specified output and displays it in the app.
}
\examples{
coronaaus::covid_title(type = "map")

}
