% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{An Application for Coronavirus in Australia}
\usage{
launch_app()
}
\description{
This function launches an application which was made to allow users to easily explore the
coronavirus data in Australia. The application contains a number of tabs including; an about tab,
a how tab, an example tab, a map of confirmed cases, a plot of confirmed cases, a comparison tab and a tab for sources.

There are no inputs required as shown in the example below, and the output returned is the launch of the application.
}
\examples{
coronaaus::launch_app()

}
