% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectize_input.R
\name{selectize_input}
\alias{selectize_input}
\title{Simplifying the UI of the Application}
\usage{
selectize_input(id, label = id, choices = choices, multiple = multiple)
}
\description{
This function simplifies the user interface of the app. It automatically generates a selectize input or a slider input
for the variables stored as \code{vars} (depending on the type of variable). For example, for \code{date} the input generated is a slider,
whereas for \code{type} the variable generated is a select input.
}
\examples{
library(coronaaus)
pmap(vars, coronaaus::selectize_input())

}
